WASIp1: Phase 1 correctness fixes — prestat, types, seek, poll, fst_flags
Lifts 16 wasi-testsuite fixtures from skip.json (38/72 active passes,
up from 22). Addresses the audit's correctness gaps plus one bug the
harness surfaced (the leading-slash prestat name).

- fd_prestat_get / fd_prestat_dir_name (FsFdStat.cs):
  Strip the internally-canonicalized leading "/" from the directory
  name before reporting it. wasi-libc's open_scratch_directory
  matches the prestat name bytewise against an embedder-supplied
  string like "fs-tests.dir"; "/fs-tests.dir" never matched, which
  was responsible for 41 of the 50 baseline failures. Also write
  exactly pr_name_len bytes (no nul) so guests sized to that count
  see clean buffers.

- fd_pread / fd_pwrite / fd_advise / fd_allocate / fd_filestat_set_size
  (FileSystem.cs):
  wasm i64 args land in the binding dispatcher as System.Int64 and
  can't auto-coerce to System.UInt64 (= filesize). Take them as
  `long` and cast to filesize once inside the lambda. Without this,
  six C tests trapped at the binding boundary with
  TargetInvocationException → ArgumentException.

- fd_seek (FsFd.cs):
  Spec slot is *newoffset: filesize (u64), but the implementation
  was writing only 4 bytes via WriteInt32 — a guest reading the
  full 8 bytes saw our int plus 4 bytes of stale linear-memory
  contents. Switch to WriteInt64.

- poll_oneoff clock subscriptions (Poll.cs):
  Previous code used DateTime.UtcNow.Ticks (100ns ticks since 0001)
  as "now" and treated the guest-supplied timeout as nanoseconds —
  absolute (SubscriptionClockAbstime) timeouts were nonsense and
  clock_id was ignored entirely. Compute `now` per subscription
  from the named clock_id (Realtime/Monotonic/CPU-time), in
  nanoseconds, and round delays up to the nearest ms.

- poll_oneoff write-readiness (Poll.cs):
  Inverted gate `Position < Length` returned ready only when the
  file was *shorter* than the position — effectively never.
  For ordinary writable streams "ready" just means the stream
  accepts writes.

- fd_filestat_set_times / path_filestat_set_times (FsFdStat.cs):
  Per spec it is an error to specify both an explicit time and the
  matching NOW flag; previously the NOW flag silently overrode the
  explicit value. Now returns ErrNo.Inval.

skip.json reshuffled to 34 entries: 8 require Phase 2 (path_link,
path_symlink, fd_fdstat_set_*), 2 require Phase 3 (sock_*), the
rest are tagged Phase 4 for later investigation.

+++ Wacs.WASIp1/FsFd.cs
@@ -574,7 +574,11 @@ namespace Wacs.WASIp1
             fileDescriptor.Stream.Seek(newPosition, SeekOrigin.Begin);
 
             var mem = ctx.DefaultMemory;
-            mem.WriteInt32(newoffsetPtr, (int)newPosition);
+            // Spec slot is *newoffset: filesize (u64). Writing only 4 bytes
+            // would leave the upper half of the guest's 8-byte read as
+            // stale linear-memory contents, breaking any guest that reads
+            // it as u64 even on small files.
+            mem.WriteInt64(newoffsetPtr, newPosition);
 
             return ErrNo.Success;
         }
